ml_pipeline/requirements.txt
# Core Build Tools (needed for Python 3.12 compatibility)
setuptools>=68.2.2
wheel
packaging
setuptools-scm
distro

# FastAPI & Web Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# Database
pymongo==4.6.1
redis==5.0.1
motor==3.3.2

# ML & Deep Learning
tensorflow                 
numpy==1.26.4               # compatible with Python 3.12
pandas==2.1.4
scikit-learn==1.3.2
openai
chronos

# Reinforcement Learning
stable-baselines3==2.2.1
gymnasium==0.29.1

# Model Registry & Tracking
mlflow==2.9.2

# Time Series
prophet==1.1.5
statsmodels==0.14.1

# Utilities
python-dotenv==1.0.0
aiohttp==3.9.1
apscheduler==3.10.4
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

prophet
chronos-forecasting